Delete dictorderviz2.py
Aside from getting practice with PIL I made this because I thought dictionaries were inconsistent in order when called and there might be a pattern anyway.  I thought it would be more fun to write a visualization instead of just looking it up but the results were a bit of a disappointment.